+++ smartgarden-BLK-v1/include/smartgarden.h
@@ -8,17 +8,17 @@ const uint8_t VALVE_START = PinSerial::Valve_0;
 const uint8_t SPRAYER_NO = PinSerial::Sprayer - VALVE_START;
 
 // Delay for checking input sensors, call set it before smartgarden_setup()
-extern uint8 SERIAL_REG[];
-extern unsigned long smartgarden_delay;
+extern uint8_t SERIAL_REG[];
+extern uint32_t smartgarden_delay;
 
 // Current analogs signal state
-extern uint8 ANALOG_SENSOR[];
-extern uint8 HUMIDITY;
+extern uint8_t ANALOG_SENSOR[];
+extern uint8_t HUMIDITY;
 // in celcius
-extern int8 TEMPERATURE;
-extern int VALVE_CURRENT;
-extern unsigned long pompa_nyala_sejak;
-extern unsigned long pompa_mati_sampai;
+extern int8_t TEMPERATURE;
+extern int8_t VALVE_CURRENT;
+extern uint32_t pompa_nyala_sejak;
+extern uint32_t pompa_mati_sampai;
 
 // Setup pin and start ticker
 void smartgarden_setup();
